# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Api Documentation",
    "version": "1.0",
    "title": "Api Documentation",
    "termsOfService": "urn:tos",
    "contact": {},
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0"
    }
  },
  "host": "localhost:18084",
  "basePath": "/",
...
  "paths": {
    "/catalog/software/": {
      "get": {
        "tags": [
          "catalog-controller"
        ],
        "summary": "software catalog list(all)",
        "description": "software catalog 리스트 불러오기",
        "operationId": "getCatalogListUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/CatalogDTO"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "catalog-controller"
        ],
        "summary": "software catalog insert",
        "description": "software catalog 등록",
        "operationId": "createCatalogUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "catalogRefData[0].catalogRefIdx",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "catalogRefData[0].catalogIdx",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "catalogRefData[0].refernectIdx",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "catalogRefData[0].referenceValue",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "catalogRefData[0].referenceDescription",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "catalogRefData[0].referenceType",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "catalogIdx",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "catalogTitle",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "catalogDescription",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "catalogSummary",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "catalogIcon",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "catalogCategory",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/CatalogDTO"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "catalog-controller"
        ],
        "summary": "software catalog update",
        "description": "software catalog 수정",
        "operationId": "updateCatalogUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "catalogRefData[0].catalogRefIdx",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "catalogRefData[0].catalogIdx",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "catalogRefData[0].refernectIdx",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "catalogRefData[0].referenceValue",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "catalogRefData[0].referenceDescription",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "catalogRefData[0].referenceType",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "catalogIdx",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "catalogTitle",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "catalogDescription",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "catalogSummary",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "catalogIcon",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "catalogCategory",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/CatalogDTO"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/catalog/software/{catalogIdx}": {
      "get": {
        "tags": [
          "catalog-controller"
        ],
        "summary": "software catalog detail",
        "description": "software catalog 내용 확인(연결된 정보들까지)",
        "operationId": "getCatalogDetailUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "catalogIdx",
            "in": "path",
            "description": "catalogIdx",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/CatalogDTO"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "catalog-controller"
        ],
        "summary": "software catalog delete",
        "description": "software catalog 삭제",
        "operationId": "deleteCatalogUsingDELETE",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "catalogIdx",
            "in": "path",
            "description": "catalogIdx",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "boolean"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "204": {
            "description": "No Content"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "CatalogDTO": {
      "type": "object",
      "properties": {
        "catalogCategory": {
          "type": "string"
        },
        "catalogDescription": {
          "type": "string"
        },
        "catalogIcon": {
          "type": "string"
        },
        "catalogIdx": {
          "type": "integer",
          "format": "int32"
        },
        "catalogRefData": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/CatalogRefDTO"
          }
        },
        "catalogSummary": {
          "type": "string"
        },
        "catalogTitle": {
          "type": "string"
        }
      }
    },
...
    "CatalogRefDTO": {
      "type": "object",
      "properties": {
        "catalogIdx": {
          "type": "integer",
          "format": "int32"
        },
        "catalogRefIdx": {
          "type": "integer",
          "format": "int32"
        },
        "referenceDescription": {
          "type": "string"
        },
        "referenceType": {
          "type": "string"
        },
        "referenceValue": {
          "type": "string"
        },
        "refernectIdx": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
...
  }
}